working on code refactoring
