+++ admin-ui/app/home/index.tsx
@@ -1,0 +1,11 @@
+
+
+const index = () => {
+   return (
+     <>this is index page</>
+    )
+  }
+  
+  export default index;
+
+
