+++ admin-ui/app/home/layout.tsx
@@ -1,0 +1,10 @@
+"use client";
+
+
+export default function Layout({ children }) {
+  return (
+    <>
+      <main>{children}</main>
+    </>
+  )
+}
