+++ admin-ui/app/home/page.tsx
@@ -1,0 +1,12 @@
+
+"use client";
+
+const page = () => {
+  
+  return (
+    <>this is home page</>
+  )
+}
+
+export default page
+
