+++ admin-ui/app/project/index.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import ListProjectComponent from "./create/ListProjectComponent";
+import ListProjectComponent from "./list";
 
 const HomeComponent = () => {
   return <ListProjectComponent />;
