+++ admin-ui/builder/index.ts
@@ -1,0 +1,1 @@
+console.log("This is index file");
